Fix #1770: Fix __scala_== with ScalaNumbers. (#1805)
The implementation of `java.lang.Number.__scala_==` was correct in
spirit, but failed in practice because of quirks of the linker and
`java.lang.Object.__scala_==`.

Normally, `Object.__scala_==` should be implemented as
`this.equals(that)`. For performance reasons, it is instead
implemented as `this eq that`, with the understanding that the
linker will patch the vtable of any class that overrides `equals`
so that `__scala_==` refers to that `equals`.

The issue is that this patching does not apply to super calls of the
form `super.__scala_==`, making any such call incorrect.

We fix the issue by directly calling `this.equals(that)` instead of
`super.__scala_==`, since the former would be the reasonable
implementation of `Object.__scala_==`.