Find the nativelib on the classpath in a more reliable way.
The tools need to know which of the classpath entries is the core
`nativelib`, because they need to extract `.c` files from it.
Previously, this was done using dirty `.contains()` checks on the
absolute path name. This is not reliable, as it can detect files
that happen to be in the workspace if the path to the workspace
contains the strings `"scala-native"` and `"nativelib"`. At the
same time, it was too strict, as it would not find the `nativelib`
if it did not come from a cache for Ivy or Coursier.

In this commit, we reliably use the Attributes of the
`fullClasspath` entries to find the one coming from a `ModuleID`
with organization `org.scala-native` and artifact name `nativelib`.