Redesign the build to use dependsOn instead of publishLocal. (#1810)
Previously, the build made heavy use of `publishLocal` to build and
publish its artifacts, which were then used by other artifacts with
`libraryDependencies`. This caused a lot of time spent in
rebuilding and republishing artifacts all the time, and required
the use of a special `rebuild` command to clean things up.

In this commit, we turn all those implicit dependencies into actual
internal dependencies, using `dependsOn`. This allows sbt to
correctly track internal dependencies itself, and rebuild when
needed.

The only tricky aspects is that some artifacts must absolutely be
exposed in their .jar form to the rest of the build:

* the compiler plugin, because scalac can only load compiler
  plugins from jars
* `nativelib`, because the linker assumes that it is a jar (it
  opens it to extract native C code from it)
* some other core libraries, because they manipulate their
  `products` (e.g., removing generated .class files)

We simply use `exportJars := true` to enforce that.

In order to cancel the `libraryDependencies` added by the
`ScalaNativePlugin` on the artifacts built in the build itself, we
create a custom `MyScalaNativePlugin` on top of it, which removes
them. Instead of relying on those automatic `libraryDependencies`,
we must explicitly tell all the required dependencies (and only
those) as `dependsOn` clauses.

---

The above changes mean that most sbt commands, like `tests/test`,
will not test the artifacts in their published form anymore.
Instead, they will test the jars produced within the build. The
published artifact are still tested through the scripted tests of
the sbt plugin, so we are not losing any test coverage in the
process.

---

In the process, we cleaned up several stale or dubious internal
dependencies. This eventually led to rethink what classpaths are
passed down to the `tools/test` task, so that was cleaned up as
well.

---

The `rebuild` command is not necessary anymore. `clean` should be
used instead.

The `dirty-rebuild` command is not necessary anymore. Simply do
*nothing* instead.

The `cleanCache` and `cleanLocal` commands, provided by
sbt-dirty-money, and which used to be used by `rebuild`, are not
necessary anymore either. Therefore, we have completely removed
sbt-dirty-money. If a developer still wants to use these commands
for their own purposes, they can globally install sbt-dirty-money,
as suggested in their readme.

---

We also find the nativelib on the classpath in a more reliable way.

The tools need to know which of the classpath entries is the core
`nativelib`, because they need to extract `.c` files from it.
Previously, this was done using dirty `.contains()` checks on the
absolute path name. This is not reliable, as it can detect files
that happen to be in the workspace if the path to the workspace
contains the strings `"scala-native"` and `"nativelib"`. At the
same time, it was too strict, as it would not find the `nativelib`
if it did not come from a cache for Ivy or Coursier.

In this commit, we reliably use the Attributes of the
`fullClasspath` entries to find the one coming from a `ModuleID`
with organization `org.scala-native` and artifact name `nativelib`.