Give testinterface ComRunner better receive() handling, remove timeout and associated bug (#1813)
This PR addresses code on the SBT side doing something ill-designed
which presents an inscrutable socket error to the end user. It does
not directly address issues in the SN binary.

The idea is to reduce protocol or handshake errors, so that errors
on the SN binary side become more evident.

All in receive():

- Mark is never set or used, so deleted.
- The timeout was never used, so deleted.
- Remove recursion from the readMessage() section, turn it into a
  local tailrec def.

Removal of the timeout handling removes the offending finally{}
which would, after an exception, set the socket timeout on a closed
socket. This caused the Exception displayed to the user to be one
about a closed socket rather that any meager original information.