Fix #1858: User-facing documentation is now free of warnings
In order to make progress on long standing issue #1260
I needed to eliminate warnings in the Sphinx build of the user-facing
documentation.

That way, any errors that I introduce will stand out.

This PR worthwhile because without it other developers will experience
the same dis-incentive and make proper documentation unlikely.