Tidy most doc build warnings revealed by Sphinx linkcheck (#1861)
The Sphinx documentation translator has a linter, called linkcheck,
which will check that external hyperlinks are valid, in the sense of
accepting a connection.

Prior to this PR, linkcheck would generate a large number of Warnings
and re-directions.  Re-directions do not affect correctness but
slow following the links. Warnings are a concern.

This PR removes those warnings and re-directions in files in and
below `docs` directory and outside of `docs/contrib`. The latter will be
fixed in a PR to follow shortly.

Where possible, hyperlinks were converted to use https transport.
At least one totally broken link was corrected. Edits were
also made in several places to make what looked like intended links
to actually be links.

The motivation for this PR, beyond Cotton Mather's dictum, is that
I am working on a PR to add documentation to the Contributor's Guide
describing how to use Sphinx to build the Scala Native user-facing
documentation. I want to document using linkcheck to
validate external links before building the docs.

This PR clears away pages of errors and allows me to find and fix errors in
the work it interrupted. It should have the same benefit to
reviewers of my next PR.