Update author list, copyright, and version for ScalaNative.pdf
The copyright seems to never be used but I updated it just for
good form.  I also commented which of the two places where the
authorship is declared.

The changes are useful at present but my real goal is to capture
the places to change for the next official Scala Native release.
I found this intervention point in the course of other work and
would have a hard time finding it again.

All changes are draft or baseline suggestions.  I can change them
as suggested by review.