Fix #657: Give libraries a way to include native C code to be compiled. (#1637)
Previously, C source files in the nativelib's jar (i.e., in its
resources), where compiled and linked together with the Scala
Native-emitted code. That behavior was reserved for the nativelib,
and not applied to user-defined libraries.

This commit generalizes this support to all libraries. Native code
source files present anywhere on the classpath will be compiled and
linked together in the final executables.

This allows libraries to include native code in their distribution.

A followup commit should restrict the set of included files to one
specific subdirectory of the classpath entries, in order to allow
libraries to also distribute source files in their actual resources
for other purposes.