Import the testing infrastructure of Scala.js. (#1869)
This commit contains a port of Scala.js' testing interface,
replacing the existing communication mechanism between the JVM and
Native test runners.

The port comes from the state of scala-js/scala-js v1.1.1 and
scala-js/scala-js-js-envs v1.1.1. The latest major redesign of this
infrastructure in Scala.js was from
https://github.com/scala-js/scala-js/pull/3122, which contains more
information about the rationale for the changes and the design.

This change may also allow in the future to generalize the testing
mechanism in both projects, which may allow to share their code.