Fix #1801: Store the already processed byte string in Val.Chars (#1855)
Previously, the `Val.Chars` IR node contained a `String` coming
more or less directly from the source code, notably still
containing escape sequences. The CodeGen was responsible for
processing the escape sequences and emitting an escaped LLVM byte
from the string.

This was very fragile, and left in the NIR specification concerns
that are specific to Scala syntax. It was fragile enough to cause
issues like #1801.

The spec of `Val.Chars` is now changed to contain a byte string,
without any notion of escape sequences. The compiler back-end is
responsible for treat the escape sequences and creating a byte
string from the source string. The CodeGen is responsible for
encoding this byte string in the LLVM syntax.

A deserialization hack is introduced not to break backward binary
compatibility.