Fix #1669: Put private methods in a separate scope through mangling. (#1898)
As described in the issue, private methods with the same arguments
and return types may have duplicate mangled names, although they
come from different classes in the hierarchy. This causes undefined
behavior and ends up executing the wrong method.

In this commit, we add a `Scope` to the mangled signatures of
`Method` and `Field`. A Scope is either `Scope.Public` or
`Scope.Private(in: Global)`. With appropriate changes to mangling,
this allows get rid of duplicate names.