Expunge emacs backup files
Ekrich kindly pointed out to me that I had committed '*.scala~' files
and wondered what & why. Those are emacs backup files and should never
have reached the main GitHub repository. Mea culpa!

This PR removes those files. There should be no functional change
and future developers will not have to spend time and energy wondering
what I was thinking.

The file 'ExecTest.scala.disabled' is correctly in the repository.
It contains "tier two (or three)" diagnostics which should not be
executed in every Travis CI run but which will be useful to a future
developer when tracking down bugs in the regex code. This was explained
in the original commit message but is not immediately evident
when encountering the file "on sight".

Regex code, especially test code, is notoriously hard to get right. Best to
capture and preserve it once right so that some future developer does not
spend weeks reproducing it and still getting it wrong.  Yes, there is
a cognitive overhead to having a ".disabled" file in the directory.
The benefit is worth the cost, although the benefit comes later and
the cost is paid now.