Add Scala 2.12 support (#1877)
The Scala 2.12 changelog can be found at
https://www.scala-lang.org/news/2.12.0/
This PR takes into account all the changes and adapts them for
Scala Native.

---

Lambdas

In order to keep the NIR changes to a minimum, we continue
generating a `FunctionN` class with an `apply` method for each
lambda. All the lambda captures are stored inside the class
instance and initialised via its constructor. The `apply` method
calls the generated `$anonfun` method, also extracting and
forwarding the captures.

---

NIR changes

Scala 2.12 now compiles trait methods as default methods in Java
interfaces. This means that the NIR was enhanced to support default
methods. The biggest impact of that is found in the reachability
analysis (`Reach.scala`).

---

Scala 2.11 cross-compilation

In order to keep the project compilable with Scala 2.11, a
`scala.scalanative.nscplugin.NirCompat` component is introduced in
`nscplugin`. This allows the compiler plugin to function with both
Scala 2.11 and 2.12.

In addition, some tests need to know under which Scala version they
are running. We use `sbt-buildinfo` to provide Scala version
information to the projects that need them (mainly for `unit-tests`
and `tools` tests).

---

Co-authored-by: Denys Shabalin <denys.shabalin@epfl.ch>