Fixed read value bigger than `0x7f` via `BufferedInputStream`
`read()` should return a value between `-1` and `255` and `-1` means `EOF`.

When `buf(pos)` contains value bigger than `0x7f` for example `171.toByte` and
makes `.toInt`, it converts this value to `-85` that can be regarded by enduser
as `EOF` if he uses `< 0` condition.

Unfortunately this bug hasn't got workaround on user side because when stream
contains `0xff`, the `read()` returns `-1` and it is impossibly to distinguish
the `EOF` and `0xff`.