Store source code positions in NIR (#1878)
This commit adds support for embedding source code positions inside
the NIR format, which is first step in enabling debugging in Scala
Native.

We add `Position` metadata to `Defn` and `Inst` via implicit
parameters. The positions are preserved through transformations and
code generation.

The serialization of Positions was ported from Scala.js.