Lower memory usage in CodeGen (#1979)
This commit is one of the steps in order to resolve #1976. It
changes the way the .ll files are generated. The previous
implementation had problems with large files, for example unit
tests in release full mode, where the body of the generated LLVM
code was of size ~900MB. This caused multiple OutOfMemory errors.

The new implementation uses another strategy. It generates the body
in a separate file and concatenates it with headers generated
afterward. Thanks to writing directly to the file, instead of using
StringBuilder and keeping all the data in memory, we're able to
significantly lower memory usage in CodeGen.