Tweak timings in ProcessTest to reduce chance of j-random CI failures.
ProcessTest has been showing intermittent failures for a long time.
This has been especially true when Travis CI is slow. It is evident
that the previous design & timings no longer fit the situation.

Recently merged PR #1966 allowed ProcessTest to give more informative
messages when it failed.  This PR adjusts the timings of several test
based on that improved information.

This is probably not a final iteration. It this is the first of probably
several rounds of "tweak; exercise; re-tweak".  The idea is to get to
the point where the logic of the tests is predominant, timing quirks
recede into the background, and the tests do not take excessive
wall clock time.