WIP: Annotate  & IOU-test j.u.List default methods
This PR builds upon merged PR #1937 by adding the JavaDefaultMethod
annotation to the trait

IOU: The discerning reader will notice that there is no ListTest.scala
     in this WIP.

     Before this PR can move out of WIP state the corresponding
     Scala.js ListTest.scala should be ported. Currently ScalaNative
     has no ListTest.

     Scala.js ListTest has a number of prerequisites, few of which
     are currently ported (such as TestMainBase).  I believe
     porting those items in a series of separate PRs. Tedious, but
     it will leave a better audit trail.

See Issue #1972 for background & discussion of JavaDefaultMethod
annotation.

This PR is WIP because it will fail to build until after PR #1997
is merged and this PR re-based upon it.