WIP: Annotate  & test j.u.f.Function default methods
This PR builds upon merged PR #1937 by porting Function.scala and its
corresponding test from Scala.js.

This approach was taken instead of editing the existing file for three
reasons:
  1) This keeps the code and its test consistent.

  2) The implementation in the Scala.js code looks like it allocates
     fewer memory objects.

  3) This makes the Scala Native & Scala.js code identical.

See Issue #1972 for background & discussion of JavaDefaultMethod annotation.

This PR is WIP because it will fail to build until after PR #1997
is merged and this PR re-based upon it.