WIP: Annotate & test j.u.f.Consumer default method
This PR builds upon merged PR #1937 by adding the JavaDefaultMethod
annotation to the trait and porting the test for the trait.

Note Well: Scala.js j.u.f.Consumer.scala carries a `@FunctionalInterface`
           annotation.  I did not port that over.  In another pass
	   I will have to figure out if SN needs such an annotation
	   and its implementation.

See Issue #1972 for background & discussion of JavaDefaultMethod annotation.

This PR is WIP because it will fail to build until after PR #1997
is merged and this PR re-based upon it.