Fix #1972: Implement JavaDefaultMethods on Scala 2.11 (#1997)
This commit is informed by the Scala.js PR #2141
"Fix scala-js/scala-js#1269: Add support for default methods."
https://github.com/scala-js/scala-js/pull/2141

We introduce a new annotation
`scala.scalanative.annotation.JavaDefaultMethod` to use on default
methods in the JDK implementation. On Scala 2.11, it amends the
codegen to force them to be compiled in a way that is binary
compatible with Java default methods, which allows them to be used
as such in codebases downstream.

On Scala 2.12+, the annotation has no effect, since the Scala
compiler already does the right thing for all concrete trait
methods.