Fix #2025: Remove BoxedUnitArray (#2036)
This commit removes `BoxedUnitArray` from concrete array
implementations. It follows from the fact that `Array[Unit]` is
always translated to `Array[BoxedUnit]`, which makes it an instance
of `Array[AnyRef]`.

Due to the limited information set of `Class[_]` we cannot get the
correct `Class.getComponentType`. This may lead to problems at
run-time. Only types with implemented component types were concrete
`TArray` instances have a correct `getComponentType`.
`BoxedUnitArray` was one of them, but not anymore after this
commit. This introduces a regression as we no longer can
distinguish `Array[BoxedUnit]` from other instances of
`Array[_ <: AnyRef]`. However, that is no different than from all
the other arrays of reference types.

We add a new scalalib 2.13 override in order to fix the only
existing usage of pattern matching on `Array[BoxedUnit]` in the
Scala stdlib.