Fix #2024: Use a shared Scope in the sbt plugin for all parallel tasks (#2039)
When creating a `JarDirectory`, it creates a new `FileSystem` with
a URI based on the jar path. It is acquired by the enclosing
`Scope`, and closed after that scope is closed. When building a
single application, it does not yield any problem. However, when
there are multiple modules linked into executables in parallel,
exiting one of the `Scope`s may close a `FileSystem´ that is still
used in other threads. It may result in a
`ClosedFileSystemException`.

This commit introduces a shared Scope used by tasks running as part
of the same command, and closed in sbt's `onComplete` event.

We also change the implementation of `Scope` to use an
`AtomicReference` instead of a `var`, so that `Scope` itself
becomes thread-safe.