Fix #2049: Use natural ordering for Arrays#sort with null comparator (#2050)
We re-port Arrays.scala from Scala.js to pick up the title fix.
An added benefit is that the various sorts are now stable, as
as described in the Java 8 documentation.

The implementation of the deepToString method is retained on a
"first do no harm" basis. The Scala.js origin uses JavaScript
dynamic arrays. ArrayDeque is an obvious replacement but that
might cause a performance regression.

ArraysTest.scala is ported to exercise the title fix. In addition,
the ported ArraysTest gives much better code coverage throughout.