WIP: Implement ArrayDequeTest for Collections environment
--- WIP commit message

    This submission is marked Work In Progress (WIP)
    because it depends upon files PR #2052, which is awaiting review.

    Until then, it will fail to build & execute. All tests pass
    locally.

    After that PR is merged, this work can be promoted to a full PR
    and this section deleted.

    An ArrayDequetest which works with in the  Collections Test environment
    will be used by a PR in waiting for
    CollectionsOnChecked{Collection, List}Test and
    CollectionsOnSynchronized{Collection, List}Test

--- PR commit message --

We port DequeTest.scala and re-port ArrayDequeTest.scala for to gain
the increased coverage of the Collections Test environment.

The previous version of ArrayDequeTest had an `@Ignored`ed case to
document a condition which fails on Scala Native, Issue #1694.  That case
is added to the re-ported ArrayDequeueTest to aid resolution of the
indicated issue.