Fix 1664: Correct 0.4.0 regression in POSIX strftime & a flaw in strptime.
We correct the definition of strftime so that the Scala Native "glue"
layer which handles translations of structure sizes is used. We also
correct a flaw in strptime where it could write memory outside of
that provided to it.

This PR updates and supersedes PR #1667. See that PR for discussion
and history.