Fix #1121: Allow ScalaDoc generation not to emit warnings (#2058)
When generating docs, the compiler does not contain all the regular
phases. Since we declare ordering dependencies wrt. those phases,
we get the following warning:

    dropping dependency on node with no phase object: mixin

In a build with -Xfatal-warnings, this results in an error.

This commit makes the component lists in `NativePlugin` and
`JUnitNativePlugin` dynamic, based on whether we are generating
documentation or not. This removes the spurious warnings.

Additionally, in our own build, we add new settings containing
extra `apiMappings` for Java.

Generation of ScalaDoc was disabled in the following projects:

* javalib - ends up with MalformedInputException: Input length = 1
* scalalib - ScalaDoc should not be generated
* projects using noPublish settings, eg: tests, sandbox,
  junitTestOutputsX