Update author list, copyright, and version for ScalaNative.pdf (#1863)
The copyright seems to never be used but it was updated it just for
good form.  Also there was commented which of the two places where the
authorship is declared.

The changes are useful at present but the real goal is to capture
the places to change for the next official Scala Native release.