Fix 1664: Correct 0.4.0 regression in strftime & memory spray in strptime.
We correct the definition of strftime & strptime so that the Scala Native
"glue" layer which handles translations of structure sizes is used. We also
correct a flaw in strptime where previously it could write memory outside
of that provided to it: memory spray.

This PR supersedes declined PR #2093. See that PR and the chain
it supersedes for discussion.