Correct j.i.DataInputStream static readUTF implementation.
Sébastian suggested in a review comment in PR #2153 that the
implementation of the readUTF() method be in the Object, with the class
implementation delegating to the Object implementation.

Design studies to implement that suggestion revealed a bug in
the implementation of the Object readUTF(): an usage of
`asInstanceOf[DataInputStream]` which could fail at runtime when
called with a valid argument which is not a superclass of DataInputStream.

We correct this defect and, in the process, implement the suggestion
which lead us to find it.

The correction is intended to make minimal changes to the Scala.js original
class implementation.