Concurrent Immix & Commix
This commit is rebased version of https://github.com/scala-native/scala-native/pull/1726

Well, it contains a lot of changes during rebase that allows to make it
works again on 24e0983e3c478325dacad44d5a7e4a7146c5cbe5.

This is the first step to split
https://github.com/catap/scala-native/tree/multithreading to some
trivial changes that can be reviews and tested.

Signed-off-by: Kirill A. Korinsky <kirill@korins.ky>
Co-authored-by: Lorenzo Gabriele <lorenzolespaul@gmail.com>