Remove debugging. No longer needed.
We harden the steps used to extract the clang-format version so
that it will now work on both Ubuntu 18.04 and 20.04.

PR #2177 removed a failure when GitHub changed "ubuntu-latest" to
20.04 by reverting the build version to 18.04. This PR removes
a hidden "gotcha" when the build version advances past 18.04.

Parsing output strings is always fraught with peril & fragile. This
PR is no exception.