--- Reviewer messag (#2218)
This PR has value in its own right. It will be used upon merge
in a PR to correct  Issue #2217, a macOS-only defect in
ResourceTest.scala.

--- Commit message ----
Implement Platform.isLinux method.

Platform.isLinux is added to extend the existing isMac and
isWindows methods. scala.util.Properties.isLinux is not
available in Scala 2.11.