Make it explicit that Unix FileAttributeView is provided by Posix calls.
Use a more accurate file name and location for a class which is private
in Java 8.

This PR is a follow up to and uses recently merged PR #2198. As
explained in that PR, this PR lays the ground work for a successor
to long standing PR #1609.  Fixing the problem addressed by PR #1609
still has value.