Removes magic constants from the commix sempahore length code (#2232)
Removes the magic constants from the code that limits the length of the semaphore names in commix GC

There are system dependent variables that limit the maximum length of semaphores for sem_open, both in POSIX compliant and in OSX systems. In OSX it should be limited by PSEMNAMLEN, in rest of the systems, by _POSIX_PATH_MAX. Source docs: POSIX.1-2017. See #2220 for discussion around PSEMNAMLEN.