WIP: Implement infrastructure to allow CI-only tests.
This Work In Progress (WIP) PR is presented for discussion
of the concept. The change by itself would be a "dead code"
PR and those are frowned upon.  It will need to be bundled
into either TimeTest or the `java.nio.file.attributes` PR
I have almost ready to submit (but not today).

Some tests have preconditions which must hold for the test to
make sense. We allow tests to determine if they are running in
a know Continuous Integration environment.