Fix #1921, #2250: Evolve hacktoberfest Posix socket sendto() & recvfrom()
PR #1921 provided an initial implementation of two Posix socket routines
usually used for UDP transport: sendto() & recvfrom().  We bring
recvfrom() into closer compliance with the posix standard, superseding
that PR.

We also fix Issue #2250 in order to improve the readability and
maintainability of the UdpSocketTest.scala used to verify the
named methods.