Allow compilation of GCs on Windows (#2264)
This PR adjusts the code of our GCs implementation allowing us to compile them on Windows as well as tries to deduplicate its common parts.
It's a cherry-pick of changes done in Windows develop branch on which implementation of both immix and none GC were working correctly. Due to some incompatibilities, Commix GC is possible to compile, but it goes into infinite loop onto first collect - this issue originated from incompatibilities between GNU and MSVC, leading to different sizes of packed structs containing bitfields. Further work in this matter is needed, however, the scope of this PR allows to use of Immix and None GC which allows for basic Windows support.

* Allow compiling GC on Windows
* Setup CI for testing compilation of GC in Windows