Allow for linktime conditional branching (#2206)
This commit adds a new concept to Scala Native by allowing to define values resolved at link-time. Based on that we're able to define link-time conditions allowing us to discard unused execution branches while performing class loading, before performing static reachability analysis.
When performing linking values can be resolved based on values defined in NativeConfig in form of Map[String, Any] with runtime checks limiting allowed values to non-empty strings and non-null primitives. In case if the property is not defined linking would fail. Link time property needs to be non-null literal. It also needs to define the explicit return type and keep resolved at its right-hand side in the same manner as it's done with extern defintions.

Link time conditions are based on if-else clauses. Runtime branching is replaced with link-time if the value used in the condition is annotated with @resolvedAtLinktime annotation. In the case of usage of mixed runtime and link-time resolved values, a compile-time error is being thrown.