Prepare build.sbt for eventual use by sbt 1.5.n.
PR #2058 introduced a regression in the consistent use of
the slash syntax recommended, but not enforced, by sbt at
the time of that PR & now.

As of sbt 1.5.0, sbt now gives a deprecation message for
the old "in (foo, bar)" style.  As of this PR, build.sbt
once again consistently uses sbt slash style so that
sbt 1.5.n can be used without that complication.