Improve, but not perfect, Posix pthread code quality.
We ameliorate Posix pathread code quality by removing `@name`
annotations in pthread.scala which are both unnecessary and have
not corresponding code in pthread.c.

A name is changed in both pthread.scala & pthread.c to make it parallel
similar names.

Lastly, `scalanative_pthread_once_init()` in pthread.c is simplified
to make it parallel other constant declarations, to show that it is
exactly similar to those, and to make it easier to trace & understand.