Add assertThrows to junit-runtime.
Ported from Scala.js, commit 4b9d2ed4 dated 2021-05-03
and commit c473689c dated 2021-05-06.

assertThrows was added to junit-runtime, expectThrows was merged
with it and assertThrows used in unit-tests and unit-tests-ext
was reimplemented using junit-runtime.

assertThrows imports in tests are now named explicitly in order to avoid
conflicts between junit-runtime and the assertThrows used in unit-tests
and unit-tests-ext.