Add a default signal handler used in unit-tests
Now, similarly to Bash programs, fatal signals in TestMain will
be reported as exit code 128 + n, where n is the signal value.
This way, ProcessRunner will be able to distinguish between planned
exits and fatal signals and report accordingly.

Signal handler was set only for signals that, by default, terminate
the process and two that stop it for unexpected reasons.

Special thanks to LeeTibbert for a significant documented research on
this subject.