Allow to use CommixGC on Windows (#2294)
This change fixes deadlock when using CommixGC in Windows, due to the difference in sizes of some structs, eg. GreyPacketsRef between Unix and Windows and the assumption that this struct should be equal to exactly 8-bytes. Such difference was caused by not portable enough code - MSVC does not support bitfields inside packed structs as expected - it seems that on the whole size of the variable is counted into the size of the struct, or some additional alignment was added.

To handle this issue, usages of 24-bit fields based on uint32_t were replaced with a dedicated UInt24 struct which should fully portable across platforms.
Additionally, there were added compile-time assertions guarding for the correct size of memory-layout fragile structs.

* Fix CommixGC: Use dedicated UInt24 struct instead of bitfields, add struct size assertions
* Add Commix GC and none GC to CI tests matrix