Ignore no longer existing objects when inheriting Interflow state (#2344)
This PR fixes issue leading to failures of release builds when using linktime resolved properties.
When resolving link-time property call to synthesised method is replaced with Op.Copy returning a resolved value.
When inlining such a call, the owner in which given property is defined (eg. scalanative.LinktimeInfo$) is never used and might be removed from the given block state. However, we don't check if a given instance is defined in both current and inherited state. This might lead to unhandled NullPointerException and failure of the build.

This PR does also contain fixes for other issues leading to build failure. Reach.addMissing now creates Path using the dedicated method for URIs, instead of passing String from URI.getPath, on Windows such usage will lead to an Invalid Argument exception.