Add a default signal handler to `TestMain` (#2338)
* Add a default signal handler used in unit-tests

Now, similarly to Bash programs, fatal signals in TestMain will
be reported as exit code 128 + n, where n is the signal value.
This way, ProcessRunner will be able to distinguish between planned
exits and fatal signals and report accordingly.

Signal handler was set only for signals that, by default, terminate
the process and two that stop it for unexpected reasons.

Special thanks to LeeTibbert for a significant documented research on
this subject.

* Change SignalConfig from C to Scala

Windows support was also added and caught signals were adjusted to match
the clib and posixlib implementations.

* Add stack trace printing to default handler

To properly work in a signal handler, stack trace printing had to be
reimplemented using only async-signal-safe methods.

* Refactor symbol formatting as to not repeat code

Async-signal-safe version of symbol formatting function was added to
nativelib. Now, javalib's StackTraceElement formatting uses that
async-signal-safe function to implement its own.