Skip linking on subsequent nativeLink calls in SN sbt plugin (#2389)
* Skip build if no files were modified since the last run.

Added file information caching based on the SHA-1 content hash of
a file. If a file was changed since the last nativeLink, the project is
rebuild and new file information about the classpath and output is
cached. This allows the plugin to skip the building process if no files
were modified, resulting in quicker subsequent runs.

* Address reviewer concerns