Remove performance regression introduced with the previous boxing setup
This also fixes problems concerning changing the view of
MappedByteBuffer. Previously, contents of the mapping would be lost if
the view changed, as the data from the pointer would be copied into an
array, and the original MappedByteBuffer would be finalised. Now,
a diffrent structure (MappedByteBufferData) is encapsulated and passed
by every MappedByteBuffer.