Support PathMatching via Globs (#2559)
Implements FileSystems.getDefault().getPathMatcher(pattern), with pattern of type "glob:...".

Instead of translating to regex, Glob matching was implemented from scratch. This is for multiple reasons - for one I believe it will be easier to understand and maintain, than doing a character based translation to regex. Additionally glob exclusive optimizations could be introduced - here we count characters and dividers (‚/’) left, more easily filtering out the states when we know we will not be able to fulfill globs.