Use `-O2` in debug mode for C files
Debug builds don't inline anything in the GC code.
Since the C compilation is done only once and then the object files are
reused in successive compilations, and since the C compiler is
faster than the Scala compiler, we can use the `-O2` compile flag
to compile the C files. This makes the GC faster which can improve
the runtime speed of debug builds.
This can be particularly useful for tests which are notoriously slow
to execute in Scala Native and are almost always executed in debug mode.
I profiled some applications in debug mode and found that 9% of the time
was spent in `ObjectMeta_SetAllocated` which is a simple assignment of a byte.
5% of stacks were falling in `MathUtils_RoundToNextMultiple` which is a
rounding function which is implemented using `mul` and `div` in `-O0` mode
but gets inlined to simple bit shifts and `xor`s.
I think this is an interesting tradeoff to consider.