Fix #3222, 3221: A few javalib stream changes (#3268)
Provide an evolution of Scala Native support for the Java Stream API.

This evolution is provided to verify correctness and as the basis for further work.
Parallel streams, performance, and huge streams all need addressing.

All streams are sequential.

Almost all of the (Object) Steam class is now implemented.

Almost all of the DoubleStream class is implemented.

IntStream and LongStream are not implemented. They
can be implemented once Stream and DoubleStream have
been shown to be correct and robust.

These class and methods in other classes which depend upon these
will throw an UnsupportedOperationException.

Array, Collection, and Optional stream methods are implemented.

StreamSupport is implemented.

The Collector and Collectors classes are implemented.

Collectors are quite useful on Scala 3. Extensive tests
demonstrate how to use them idiomatically with lambda expressions.
Collectors can be made to work with Scala 2.13 with explicit classes: e.g. Consumer, et cetera.
The declarations necessary to use Collectors with Scala 2.12 are likely to make the effort not
worth the reward.
An improved algorithm for mapMulti methods is implemented.

Except where noted in this announcement, all Java 8 through 20 methods are implemented.