Delibrately deal with thread_local and counters
Obviously on multi-core machines creating handlers, using an atomic counter
for the handlers is a must.

Otherwise, note that while variables can be thread-local and re-read in code,
their addresses (after receiving them from the OS) can be stored in registers.
This means that, upon suspend/restore on another thread, we can still the
variable in the old thread!
We must be careful not to use thread-local variables across suspend/resume
boundary. For `handlers` chain, the simple solution is to just not inline
`handlers_` library operations (and so the function boundary is maintained,
and no `handlers_` library functions are suspended in the middle).