Fix: do not stack overflow when exception occurs in Backtrace.decodeFileline
When analyzing the stack trace, if the entry "currentStackTrace" appears more than once, it indicates that a Throwable object
was generated within the same "currentStackTrace" method, causing recursive calls. This recursive behavior can lead to an
infinite loop, ultimately resulting in a stack overflow exception.

To prevent excessive recursion, it's preferable to minimize multiple consecutive calls to the "currentStackTrace" method.

The "currentStackTrace" process is straightforward and typically doesn't trigger exceptions.
On the other hand, the "BackTrace.decodeFileline" is intricate, and if an exception thrown, it's likely to originate from that method.

Consequently, to mitigate the risk of cascading recursive exceptions,
skip executing "BackTrace.decodeFileline" if "currentStackTrace" is already being invoked recursively.