Implement delimcc for x86-linux
The implementation mostly just adds the Assembly part. Due to x86 passing most things
through stack we have to be more careful with cleaning up after ourselves with the parameters.
Furthermore, it is not possible to keep the "stack pointer offset = 8" claim without
more complicated Assembly code, so I just opted to make it special for x86.