Implement delimited continuations primitives (#3286)
* Delimcc implementation for arm64
* Use ObjectArray for object boxes
* Implement the same thing for amd64 linux
* Move continuation allocations to Scala Native
Theoretically this should allow us to fully control long-living allocations
and allow GCs to scan stored stack blobs.

The idea is
<--[ __cont_resume_impl ][ resumed stack ][ cont_resume ] ... --|
- _lh_resume_entry reserves the resumed stack's amount on the stack, then jumps to __cont_resume_impl.
- __cont_resume_impl sets up the resumed stack (above it), then longjmps (upwards!) to it.

* Store allocated blobs inside an explicit Continuation object
* Add support for arm64 darwin
* Handle exceptions with tunneling Try
* Sync implementation with sysv64
* Delibrately deal with thread_local and counters
Obviously on multi-core machines creating handlers, using an atomic counter
for the handlers is a must.

Otherwise, note that while variables can be thread-local and re-read in code,
their addresses (after receiving them from the OS) can be stored in registers.
This means that, upon suspend/restore on another thread, we can still the
variable in the old thread!
We must be careful not to use thread-local variables across suspend/resume
boundary. For `handlers` chain, the simple solution is to just not inline
`handlers_` library operations (and so the function boundary is maintained,
and no `handlers_` library functions are suspended in the middle).

* Run the suspend parameter function after setting up the jump buffer

Otherwise the function cannot resume the suspension right away, and
we are passing it an incomplete Suspension.

* Add support for x86_64-darwin

Looks like MacOS uses the same syscall convention as Linux on x86_64,
so simply allowing the constants to be defined seems to work.


* Add (not working yet) Windows support

Big hurdle: Windows exc handling forces the use of a frame pointer,
which means we are unable to support code that has `try ... catch`.
This includes the `Try` wrapper, so everything is just straight up
crashing at the moment. Will need to dig deeper.

* Add link time error for Continuations on Windows