Remove 'BackendPipeline'
The class doesn't actual represent any useful abstraction. It is merely
a namespace for a handful of related methods. This commit redefines
those methods in the namespace of `Build`, thus reducing indentation.

Because of this change, the compiler configuration and the result of
the linker have to be passed explicitly while composing compilation
steps (see the `build` method).