Improve the API and documentation of build utilities (#3525)
* Add documentation to utility methods

* Rename 'getLatestMtime' to 'getLastModifiedChild'

* Remove 'BackendPipeline'

The class doesn't actual represent any useful abstraction. It is merely
a namespace for a handful of related methods. This commit redefines
those methods in the namespace of `Build`, thus reducing indentation.

Because of this change, the compiler configuration and the result of
the linker have to be passed explicitly while composing compilation
steps (see the `build` method).

* Fix parameter name in method documentation

* Apply scalafmt

* Omit needless words in method documentation

* Improve return value docs

* Add missing method docs

* Rename 'findAndCompileNativeLibs' to 'findAndCompileNativeLibraries'