Add `pledge` and `unveil`
`pledge` is using to restrict system operations, and `unveil` to unveil
parts of a restricted filesystem view.

Both of them available at OpenBSD for years. Anyway some work to port it
to Linux is onging but it far from the end, and the best status is
Justine's work to port it into cosmopolitan libc.