improvement [docs]: remove `$`, separate comand/output, from cli snippets
resolves https://github.com/scala-native/scala-native/issues/4001

Improves quality of life for unix cli power users. It is obvious to the typical reader that these snippets are cli commands so the `$` is not of much benefit, and removing it allows users to copy paste all commands in one go instead of manually copying and pasting each individual command which is a larger benefit. Separating the command outputs into their own snippets allows for easier copy pasting without accidentally selecting the output.