Get function name from DWARF when available
`unw_get_proc_name` doesn't work on static binaries.
This PR uses DWARF to get the procedure names at runtime
when they are available.
It uses `DW_AT_linkage_name` which contains mangled names for functions.
The `main` function has an ugly `_SM2__C4main` name which is meaningless.
I can add a special case for that (or maybe we can consider removing the generated
main function from the stack traces altogether, so we don't need to think about it